Add bar behind diamond.
If a shape takes more than one aa, a bar behind it should appear, taking
the width corresponding to all the involved aa. So far only diamond is
working. Chevron is centered in the middle of the aa invloved, bar is
missing. Centering and bars are missing for all the others.