remove duplicate variant graph
This seems to plot the 'stroke', then separately plot the 'fill'.
The duplication doesn't seem necessary and the result is confusing on mouseovers, etc.
This change combines the CSS classes to achieve a similar style.

Also, this changes the interpolation from linear to step since that
seems more appropriate for discrete data points?