use left/main panels to properly align columns
there was some inconsistency in how the left column (categories) and
the main column (header and features) were applied. It meant that the
column edges didn't line up perfectly. This simplifies things by using
a 'leftpanel' and 'mainpanel' container, the margins of which are only
specified in a single place in the CSS.

Also, the contents should now be able to happily fill up to 100% width of
the container which may help when attempting to allow flexibility in
changing the width.